requirements.txt
xpander_sdk
openai
python-dotenv 